refactor getmessagesbyid
